MediaFocusControl: Do not grant audio focus if notification stream is 0
Previously audio focus would be granted regardless of notification volume, meaning
that if notifications are muted the media stream would still be dimished for focus.
By checking if STREAM_NOTIFICATION is 0 and if a notification stream is
requesting focus, this no longer an issue.

PS1: Initial implement
PS2: Better implement in MediaFocusControl
PS3: Whitespace

Change-Id: Ie10a64e58ce77174a3bbdeaa936646d40c24404f
Signed-off-by: Evan Anderson <evan1124@gmail.com>